+++ praktikumsplaner-frontend/frontend/src/components/praktikumsplaetze/Meldung/StudiensemesterSelect.vue
